+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fix thread safety in process code #43
+- Fix process package build on AIX #54
 
 ## [0.4.4]